+++ Source/Core/Core/WiiRoot.cpp
@@ -11,11 +11,11 @@
 #include "Common/Logging/Log.h"
 #include "Common/NandPaths.h"
 #include "Common/StringUtil.h"
-#include "Common/SysConf.h"
 #include "Core/ConfigManager.h"
 #include "Core/IOS/IOS.h"
 #include "Core/Movie.h"
 #include "Core/NetPlayClient.h"
+#include "Core/SysConf.h"
 
 namespace Core
 {
