+++ Source/Core/DolphinWX/GameListCtrl.cpp
@@ -39,7 +39,6 @@
 #include "Common/CommonTypes.h"
 #include "Common/FileUtil.h"
 #include "Common/StringUtil.h"
-#include "Common/SysConf.h"
 #include "Common/Thread.h"
 #include "Core/Boot/Boot.h"
 #include "Core/Config/NetplaySettings.h"
@@ -48,6 +47,7 @@
 #include "Core/HW/DVD/DVDInterface.h"
 #include "Core/HW/WiiSaveCrypted.h"
 #include "Core/Movie.h"
+#include "Core/SysConf.h"
 #include "Core/TitleDatabase.h"
 #include "Core/WiiUtils.h"
 #include "DiscIO/Blob.h"
